requirements.txt
streamlit
python-nmap
psycopg2-binary
pandas
plotly
pdfkit
jinja2
SQLAlchemy
weasyprint